+++ integrations/anthropic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [integrations/anthropic-v2.6.1] - 2025-06-16
+
+### 🌀 Miscellaneous
+
+- Fix: `AnthropicChatGenerator` now properly can call tools that have no arguments when streaming is enabled (#1950)
+
 ## [integrations/anthropic-v2.6.0] - 2025-06-13
 
 ### 🐛 Bug Fixes